apps/python-backend/pyproject.toml
[project]
name = "financial_planner"
authors = [
  { name = "financial_planner", email = "king.arthur@camelot.bt" }
]
description = "Financial Planning Tools"
readme = "../../README.md"
dynamic = ["version"]
classifiers = [
    "Programming Language :: Python :: 3",
]
requires-python = ">=3.9"
dependencies = [
  "python-multipart",
  "uvicorn>=0.34.0",
  "pyyaml>=6.0.2",
  "fastapi>=0.115.7",
]

[project.urls]
Documentation = "https://jcmullwh.github.io/financial_planner/"
Source = "https://github.com/jcmullwh/financial_planner"

[tool.mypy]
python_version = "3.9"
files = ["apps/python-backend/src/financial_planner"]
mypy_path = "apps/python-backend/src"
disallow_untyped_defs = true
disallow_any_unimported = true
no_implicit_optional = true
check_untyped_defs = true
warn_return_any = true
warn_unused_ignores = true
show_error_codes = true

[tool.ruff]
target-version = "py39"
line-length = 120

[tool.ruff.lint]
extend-select = ["I"]
select = [
    "A",
    "ARG",
    "B",
    "C",
    "DTZ",
    "E",
    "EM",
    "F",
    "FBT",
    "ICN",
    "ISC",
    "N",
    "PLC",
    "PLE",
    "PLR",
    "PLW",
    "Q",
    "RUF",
    "S",
    "T",
    "TID",
    "UP",
    "W",
    "YTT"
]
ignore = [
    "B027",
    "FBT003",
    "S105",
    "S106",
    "S107",
    "C901",
    "PLR0911",
    "PLR0912",
    "PLR0913",
    "PLR0915",
    "T201"
]

[tool.ruff.lint.per-file-ignores]
"apps/python-backend/tests/*" = ["S101","PLR2004","E712"]

[tool.pytest.ini_options]
addopts = "--cov=apps/python-backend/src/financial_planner --cov-report=term-missing"
asyncio_mode = "strict"

[tool.pdm.dev-dependencies]
test = [
  "pytest>=8.2.0",
  "pytest-cov>=5.0.0",
  "pytest-asyncio",
  "types-PyYAML>=6.0"
]
docs = [
  "mkdocs>=1.6.0",
  "mkdocs-material>=9.5.21",
  "mkdocstrings[python]>=0.25.1"
]
lint = [
  "ruff>=0.8.5",
  "mypy>=1.10.0",
  "deptry>=0.16.1",
  "pre-commit>=3.7.0",
  "tox>=4.15.0"
]

[tool.pdm.scripts]
test = "pytest"
test-cov-xml = "pytest --cov-report=xml"
lint = "python scripts/lint.py"
lint-check = "python apps/python-backend/scripts/lint-check.py"
docs-serve = "mkdocs serve"
docs-build = "mkdocs build"
backend = "uvicorn financial_planner.backend_api:app --reload"
run-frontend = "cd apps/frontend/financial-planner-frontend && npm run start"

[tool.pdm.build]
package-dir = "src"

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"
